Rename `FieldTypes2` to `FieldTypes`
